refactor(backend): isolate Codex auth policy
